add brief documentation on the func signatures of each file
