+++ Faust_Interpreter_Test_Plugin/FaustRuntime/AudioInput/audio_inputs.h
@@ -1,3 +1,13 @@
+/*******************************************************************************
+This header file defines the structure for managing different audio input 
+generators for a DSP effects. 
+The file includes the declaration of available audio inputs and provides the 
+mechanism to retrieve and create these audio inputs by ID. 
+The list of available audio inputs is defined using macros, and each input 
+corresponds to a specific audio generator or effect (e.g., Sine wave, Wind, 
+Rain, etc.).
+*******************************************************************************/
+
 #ifndef AUDIO_INPUTS_H
 #define AUDIO_INPUTS_H
 
@@ -28,8 +38,10 @@ typedef std::vector<std::pair<int, const wchar_t*>> audioInputList;
 
 namespace AudioInputs{
 
+    /* Function to retrieve the full list of available audio inputs, typically returning an audioInputList.*/
     audioInputList getAudioInputList();
 
+    /* Function that returns a new dsp instance. The caller is responsible for destroying the dsp object on time. */
     dsp* createAudioInputById(int id);
 }
