+++ Faust_Interpreter_Test_Plugin/FaustRuntime/Parameters/myMapUI.h
@@ -5,6 +5,7 @@
 #include "parameter.h"
 #include <unordered_map>
 
+// static counter to increment for each new parameter. Has to be reset on each reseting. 
 static int ctr{0};
 
 struct myMapUI : public MapUI
@@ -20,94 +21,48 @@ struct myMapUI : public MapUI
         double step;
         
         itemInfo(std::string t, std::string l, int idx, double i, double mn, double mx, double s)
-        :type(t), label(l), index(idx), init(i), fmin(mn), fmax(mx), step(s)
+            :type(t), label(l), index(idx), init(i), fmin(mn), fmax(mx), step(s)
         {}
     };
 
-    ~myMapUI() 
-    {
-        resetStaticCounter();
-    }
-
+    // controls are a map of (zone ptrs, itemInfo) pairs.
     std::unordered_map<const FAUSTFLOAT*,itemInfo> controls;
 
-    void fillShortNames()
-    {
-        for (const auto& it : fShortnameZoneMap) {
-            const auto& shortName = it.first;
-            const FAUSTFLOAT* zone = it.second;
+    ~myMapUI();
 
-            auto iter = controls.find(zone);  // only access existing item
-            if (iter != controls.end()) {
-                iter->second.shortname = shortName; // assign shortname
-            }
-        }
-    }
+    /* function that make use of fShortnameZoneMap to fill in the contros the shortName of each param. */
+    void fillShortNames();
 
-    void addParameter(FAUSTFLOAT* zone, std::string type, const char* label, int index, double init = -DBL_MAX, double fmin = -DBL_MAX, double fmax = -DBL_MAX, double step = -DBL_MAX) {
-        
-        itemInfo item(type, std::string(label), index, init, fmin, fmax, step);
-        controls.emplace(zone, std::move(item));
-    }
-
-    void resetStaticCounter()
-    {
-        ctr = 0;
-    }
-
-    void clearParameters()
-    {
-        controls.clear();
-        fLabelZoneMap.clear();
-        fShortnameZoneMap.clear();
-        fPathZoneMap.clear();
-        resetStaticCounter();
-    }
+    // function that moves a new parameter into the controls structure.
+    void addParameter(FAUSTFLOAT* zone, std::string type, const char* label, int index, double init = -DBL_MAX, double fmin = -DBL_MAX, double fmax = -DBL_MAX, double step = -DBL_MAX);
+    
+    // function to reset the static counter that counts the num of params.
+    void resetStaticCounter();
+    
+    /* function that clears all the internal structures, and resets the static counter. */
+    void clearParameters();
 
     // -- active widgets
-    void addButton(const char* label, FAUSTFLOAT* zone)
-    {
-        MapUI::addZoneLabel(label, zone);
-        addParameter(zone, "button", label, ctr++);
-    }
-    void addCheckButton(const char* label, FAUSTFLOAT* zone)
-    {
-        addZoneLabel(label, zone);
-        addParameter(zone, "checkbox", label, ctr++);
-    }
-    void addVerticalSlider(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT init, FAUSTFLOAT fmin, FAUSTFLOAT fmax, FAUSTFLOAT step)
-    {
-        addZoneLabel(label, zone);
-        addParameter(zone, "slider", label, ctr++, init, fmin, fmax, step);
-    }
-    void addHorizontalSlider(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT init, FAUSTFLOAT fmin, FAUSTFLOAT fmax, FAUSTFLOAT step)
-    {
-        addZoneLabel(label, zone);
-        addParameter(zone, "slider", label, ctr++, init, fmin, fmax, step);
-    }
-    void addNumEntry(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT init, FAUSTFLOAT fmin, FAUSTFLOAT fmax, FAUSTFLOAT step)
-    {
-        addZoneLabel(label, zone);
-        addParameter(zone, "slider", label, ctr++, init, fmin, fmax, step);
-    }
+    void addButton(const char* label, FAUSTFLOAT* zone);
+
+    void addCheckButton(const char* label, FAUSTFLOAT* zone);
+    
+    void addVerticalSlider(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT init, FAUSTFLOAT fmin, FAUSTFLOAT fmax, FAUSTFLOAT step);
+    
+    void addHorizontalSlider(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT init, FAUSTFLOAT fmin, FAUSTFLOAT fmax, FAUSTFLOAT step);
+
+    void addNumEntry(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT init, FAUSTFLOAT fmin, FAUSTFLOAT fmax, FAUSTFLOAT step);
     
     // -- passive widgets
-    void addHorizontalBargraph(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT fmin, FAUSTFLOAT fmax)
-    {
-        addZoneLabel(label, zone);
-        addParameter(zone, "bargraph", label, ctr++, -DBL_MAX, fmin, fmax, -DBL_MAX);
-    }
-    void addVerticalBargraph(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT fmin, FAUSTFLOAT fmax)
-    {
-        addZoneLabel(label, zone);
-        addParameter(zone, "bargraph", label, ctr++, -DBL_MAX, fmin, fmax, -DBL_MAX);
-    }
+    void addHorizontalBargraph(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT fmin, FAUSTFLOAT fmax);
+    
+    void addVerticalBargraph(const char* label, FAUSTFLOAT* zone, FAUSTFLOAT fmin, FAUSTFLOAT fmax);
 
     // -- soundfiles
-    // virtual void addSoundfile(const char* label, const char* filename, Soundfile** sf_zone) {}
+    // virtual void addSoundfile(const char* label, const char* filename, Soundfile** sf_zone);
     
     // -- metadata declarations
-    // virtual void declare(FAUSTFLOAT* zone, const char* key, const char* val){}
+    // virtual void declare(FAUSTFLOAT* zone, const char* key, const char* val);
 };
 
 #endif
