+++ Faust_Interpreter_Test_Plugin/FaustRuntime/PluginTemplate/abstract.h
@@ -11,14 +11,19 @@ typedef uint32_t AkUInt32;
 class AbstractPlugin{
 public:
 
+    /* Constructor */
     AbstractPlugin(PluginConfiguration&, ParameterList&, InterpreterWrapper&);
 
+    /* Destructor */
     virtual ~AbstractPlugin();
     
+    /* Function that resets the plugin to its initial state. */
     virtual void reset();
 
+    /* Function to allow any type of preprocessing before audio streaming commence. */
     virtual void setup();    
 
+    /* Pure virtual audio callback function. Is implemented differently on the children classes (source, effect). */
     virtual void callback(std::vector<FAUSTFLOAT*>&, const AkUInt32)=0;
     
 protected:
