+++ Faust_Interpreter_Test_Plugin/FaustRuntime/PluginUtils/syscall.h
@@ -5,8 +5,10 @@
 
 namespace SysCall{
     
+    /* Function that is used to get the Wwise env path. */
     std::string getEnvVar(const std::string& varName);
 
+    /* Function used to call system commands (i.e. faust --dspdir). */
     std::string execCommand(const std::string& cmd);
 
 }
